test(conversation): split the thread suite, and say what a reader sees
The renderer's suite was over the 300-line limit, and it was two questions in
one file: what a single message draws, and what the thread does around it
(which trace a turn offers, which box it scrolls). It is those two files now,
over a shared rig — the `vi.mock` calls stay per file, since they are hoisted,
but the rendering does not need copying.

The flattener scenarios in `playground-conversation.feature` described a
function rather than a customer: "when the message is flattened", "then a text
part and a tool call part are produced", `tool_calls` versus `toolCalls`. They
now say what a reader sees — the same tool call whichever dialect the provider
used, prose and a tool call in the order written, a spoken reply shown once
with its transcript — and the `@scenario` annotations follow them, so all 35
scenarios stay bound.

Two prose blocks trimmed to what their scenarios actually pin: the sidebar's
collapse paragraph claimed hover expansion and no-reflow, neither of which this
PR covers, and the runtime-parameters block explained where the values are
stored on the way to explaining where the section sits.
