+++ platform/app/src/components/conversation/__tests__/ConversationThread.parts.integration.test.tsx
@@ -1,26 +1,26 @@
 /**
  * @vitest-environment jsdom
  *
- * What the shared conversation renderer actually draws.
+ * What the shared conversation renderer draws for one message.
  *
  * `flattenMessages.unit.test.ts` covers the flattening — which wire shapes
  * become which parts. This covers the half after it: given those parts, what
  * a reader sees. Together they are the path the playground runs, which is
  * `flattenMessages` then `ConversationThread` and nothing in between, so the
- * fixtures here start from raw messages rather than hand-built parts wherever
- * the scenario says the conversation was loaded from a trace.
+ * fixtures start from raw messages rather than hand-built parts.
  *
  * These are the behaviours CopilotKit dropped: it rendered text and only text,
  * so a tool call, a reasoning block and an attachment all vanished from the
  * playground even though the trace held them.
  *
+ * The thread AROUND the messages — traces, scrolling — is
+ * `ConversationThread.thread.integration.test.tsx`.
+ *
  * Spec: specs/prompts/playground-conversation.feature
- * Spec: specs/prompts/undefined-variable-banner-stability.feature
  */
-import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
-import { cleanup, render, screen } from "@testing-library/react";
+import { cleanup, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import type { ReactElement, ReactNode } from "react";
+import type { ReactNode } from "react";
 import {
   afterEach,
   beforeAll,
@@ -30,8 +30,10 @@ import {
   it,
   vi,
 } from "vitest";
-import { ConversationThread } from "../ConversationThread";
-import { type FlattenableMessage, flattenMessages } from "../flattenMessages";
+import {
+  renderConversation,
+  traceNotYetLanded,
+} from "./conversationThreadHarness";
 
 const openTraceDetailsDrawer = vi.hoisted(() => vi.fn());
 const tracesGetByIdQuery = vi.hoisted(() => vi.fn());
@@ -65,42 +67,7 @@ vi.mock("~/features/traces-v2/components/TraceIdPeek", () => ({
   TracePreviewHoverCard: ({ children }: { children: ReactNode }) => children,
 }));
 
-const message = (msg: Record<string, unknown>) => msg as FlattenableMessage;
-
-/**
- * The playground's own configuration, so these tests answer for the surface
- * the spec is about: `PromptPlaygroundChat` renders `ConversationThread` with
- * `structuredOutput` on, the default `regular` variant (turn separators on)
- * and the default autoScroll, and adds nothing else.
- */
-function renderConversation(
-  messages: Record<string, unknown>[],
-  options: {
-    structuredOutput?: boolean;
-    labels?: { user?: string; assistant?: string };
-    roleMode?: "chat" | "scenario";
-  } = {},
-): ReturnType<typeof render> {
-  const ui: ReactElement = (
-    <ConversationThread
-      parts={flattenMessages({ messages: messages.map(message) })}
-      projectId="proj-1"
-      structuredOutput={options.structuredOutput ?? true}
-      labels={options.labels}
-      roleMode={options.roleMode ?? "chat"}
-    />
-  );
-  return render(<ChakraProvider value={defaultSystem}>{ui}</ChakraProvider>);
-}
-
-/** The trace behind a turn has landed and can be opened. */
-const traceLanded = () => ({ data: { trace_id: "trace-1" }, isError: false });
-/** The trace has not been written yet — the query is still retrying. */
-const traceNotYetLanded = () => ({ data: undefined, isError: false });
-/** The trace will never arrive: it expired, or was never written. */
-const traceGone = () => ({ data: undefined, isError: true });
-
-describe("<ConversationThread />", () => {
+describe("given a message rendered in the conversation", () => {
   // The thread scrolls the newest part into view on every render; jsdom has no
   // implementation of it.
   beforeAll(() => {
@@ -286,106 +253,6 @@ describe("<ConversationThread />", () => {
       ]);
 
       expect(screen.getByTestId("media-part-audio")).toBeInTheDocument();
-    });
-  });
-
-  describe("given a turn whose trace has landed", () => {
-    beforeEach(() => {
-      tracesGetByIdQuery.mockReturnValue(traceLanded());
-    });
-
-    /** @scenario Each completed turn offers its trace */
-    it("separates the turn with a labelled separator that opens the trace", async () => {
-      renderConversation([
-        {
-          id: "m1",
-          role: "assistant",
-          content: "done",
-          trace_id: "trace-1",
-        },
-      ]);
-
-      expect(screen.getByText("Turn 1")).toBeInTheDocument();
-
-      const separator = screen.getByRole("button", {
-        name: "View trace for turn 1",
-      });
-      await userEvent.click(separator);
-
-      expect(openTraceDetailsDrawer).toHaveBeenCalledWith({
-        traceId: "trace-1",
-      });
-    });
-  });
-
-  describe("given a turn whose trace has not landed yet", () => {
-    /** @scenario A turn whose trace has not landed yet offers no trace affordance */
-    it("leaves the separator a plain rule", () => {
-      renderConversation([
-        {
-          id: "m1",
-          role: "assistant",
-          content: "done",
-          trace_id: "trace-1",
-        },
-      ]);
-
-      expect(screen.getByText("Turn 1")).toBeInTheDocument();
-      // Advertising a trace that 404s is worse than waiting a beat for one
-      // that opens, so nothing is offered until the trace is really there.
-      expect(screen.queryByRole("button", { name: /View trace/ })).toBeNull();
-      expect(screen.queryByText("View trace")).toBeNull();
-    });
-  });
-
-  describe("given the thread is rendered beside other scrolling content", () => {
-    it("scrolls its own box rather than every ancestor", () => {
-      // jsdom implements neither, so both are observed rather than measured.
-      // Spied rather than assigned: assigning left the stub on the shared
-      // prototype for every describe after this one, and `clearAllMocks`
-      // clears calls without removing the patch.
-      const scrollTo = vi
-        .spyOn(Element.prototype, "scrollTo")
-        .mockImplementation(() => undefined);
-
-      renderConversation([{ id: "m1", role: "user", content: "hello" }]);
-
-      // `scrollIntoView` walks up and scrolls EVERY ancestor scroll container
-      // it finds. The playground puts this thread beside the prompt editor, so
-      // the thread re-mounting on a tab change dragged the editor to the top
-      // and back down again. Scrolling its own box cannot reach a sibling.
-      expect(Element.prototype.scrollIntoView).not.toHaveBeenCalled();
-      expect(scrollTo).toHaveBeenCalled();
-
-      scrollTo.mockRestore();
-    });
-  });
-
-  describe("given a conversation whose trace no longer exists", () => {
-    beforeEach(() => {
-      tracesGetByIdQuery.mockReturnValue(traceGone());
-    });
-
-    /** @scenario A prompt with a running conversation re-opens without crashing */
-    it("renders the conversation instead of throwing to the error boundary", () => {
-      // Re-opening a prompt tab replays its stored conversation, and an
-      // assistant turn in it can name a trace that expired or was never
-      // written. The affordance degrades; the transcript still renders.
-      expect(() =>
-        renderConversation([
-          { id: "m1", role: "user", content: "hello" },
-          {
-            id: "m2",
-            role: "assistant",
-            content: "hi there",
-            trace_id: "trace-gone",
-          },
-        ]),
-      ).not.toThrow();
-
-      expect(screen.getByText("hello")).toBeInTheDocument();
-      expect(screen.getByText("hi there")).toBeInTheDocument();
-      expect(screen.queryByRole("button", { name: /View trace/ })).toBeNull();
     });
   });
